move build watcher code to another internal module
connects to #41
